backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
faiss-cpu
pypdf
sentence-transformers
python-multipart

